Rewrite VectorField as VectorPlot for modularity and sensible logic
VectorField contains a Map of Point objects to vectorDrawing objects.

VectorDrawing objects are normal paperjs objects with a custom position method
(to account for the origin not being at the center of the bounding box) and
an update method that takes a new vector.

VectorField's addPoint method updates the Map, but applies a
linear transformation beforehand to place the origin in the center of the
screen and to scale by this.scaleFactor

VectorField no longer initializes with evenly spaced points, instead that
is done using the fillWithPoints method. fillWithPoints currently takes
a *function* which returns a vectorDrawing, rather than a vectorDrawing alone.
This is due to the lack of deep object cloning in javascript. This is likely
not idiomatic though.

Finally paperscript now uses a more up to date javascript parser, so newer
language features are available (yay)